Fixed bug where a player leaving the area was announced multiple times.

+++ client/lib/game/entities/non-weltmeister/network-player.js
@@ -57,14 +57,18 @@ ig.module(
 			// A player disconnected or left the area.
 			socket.on('dropPlayer-' + this.name, function() {
 				
-				// Write event to screen.
-				ig.game.events.push(player.name + " left the area.");
+				// Prevent multiple drop announcements.
+				if(!player._killed)
+				{
+					// Write event to screen.
+					ig.game.events.push(player.name + " left the area.");
 
-				// Write to chat log.
-				ig.game.chatLog.push('<div class="info"><span class="name">[' + player.name + ']</span> left the area.</div>');
+					// Write to chat log.
+					ig.game.chatLog.push('<div class="info"><span class="name">[' + player.name + ']</span> left the area.</div>');
 
-				// Free resources.
-				player.kill();
+					// Free resources.
+					player.kill();
+				}
 			});
 
 		},
